Genre.SaveGenreForArtist  and first form traits
