+++ test/TestOrganizerModel.scala
@@ -27,7 +27,7 @@ class TestOrganizerModel extends PlaySpec with OneAppPerSuite {
   val genreMethods = new GenreMethods(dbConfProvider, utilities)
   val searchSoundCloudTracks = new SearchSoundCloudTracks(dbConfProvider, utilities, trackMethods, genreMethods)
   val searchYoutubeTrack = new SearchYoutubeTracks(dbConfProvider, genreMethods, utilities, trackMethods)
-  val geographicPointMethods = new GeographicPointMethods(dbConfProvider, utilities)
+  val geographicPointMethods = new SearchGeographicPoint(dbConfProvider, utilities)
   val tariffMethods = new TariffMethods(dbConfProvider, utilities)
   val placeMethods = new PlaceMethods(dbConfProvider, geographicPointMethods, utilities)
   val addressMethods = new AddressMethods(dbConfProvider, utilities, geographicPointMethods)
@@ -85,7 +85,7 @@ class TestOrganizerModel extends PlaySpec with OneAppPerSuite {
         firstName = Option("firstName"),
         lastName = Option("lastName"),
         fullName = Option("fullName"),
-        email = Option("email"),
+        email = Option("emaill"),
         avatarURL = Option("avatarUrl"))
       whenReady(organizerMethods.saveWithAddress(OrganizerWithAddress(organizer, None)), timeout(Span(5, Seconds))) { savedOrganizer =>
         whenReady(userDAOImpl.save(user), timeout(Span(5, Seconds))) { savedUser =>
