+++ test/TestPlaceController.scala
@@ -22,7 +22,7 @@ class TestPlaceController extends PlaySpecification with Mockito {
   val genreMethods = new GenreMethods(dbConfProvider, utilities)
   val searchSoundCloudTracks = new SearchSoundCloudTracks(dbConfProvider, utilities, trackMethods, genreMethods)
   val searchYoutubeTrack = new SearchYoutubeTracks(dbConfProvider, genreMethods, utilities, trackMethods)
-  val geographicPointMethods = new GeographicPointMethods(dbConfProvider, utilities)
+  val geographicPointMethods = new SearchGeographicPoint(dbConfProvider, utilities)
   val tariffMethods = new TariffMethods(dbConfProvider, utilities)
   val placeMethods = new PlaceMethods(dbConfProvider, geographicPointMethods, utilities)
   val addressMethods = new AddressMethods(dbConfProvider, utilities, geographicPointMethods)
