+++ test/TestPlaceModel.scala
@@ -21,7 +21,7 @@ class TestPlaceModel extends PlaySpec with OneAppPerSuite {
   val injector = appBuilder.injector()
   val dbConfProvider = injector.instanceOf[DatabaseConfigProvider]
   val utilities = new Utilities
-  val geographicPointMethods = new GeographicPointMethods(dbConfProvider, utilities)
+  val geographicPointMethods = new SearchGeographicPoint(dbConfProvider, utilities)
   val trackMethods = new TrackMethods(dbConfProvider, utilities)
   val genreMethods = new GenreMethods(dbConfProvider, utilities)
   val searchSoundCloudTracks = new SearchSoundCloudTracks(dbConfProvider, utilities, trackMethods, genreMethods)
